feat: implement CH32H417 edge visual gimbal system control loop and multi-modal interaction
This commit implements the complete firmware and Python edge gateway for the CH32H417三轴稳像云台 control system, replacing the old V307 implementation with advanced control, filtering, and communication flow control.

Key changes:
1. Python Edge Gateway (识别.py):
   - Integrated SCRFD face detector (InsightFace) and MediaPipe Hands gesture classification.
   - Designed a double-EMA (Input/Output) cascade filter with a hysteresis deadzone to eliminate target pixel jitter and servo micro-vibrations.
   - Added linear extrapolation forecasting to compensate for dropped frames and maintain a 30Hz control stream.
   - Configured serial ports for com9 (Hand), com10 (Face), and com27 (MCU).
   - Handled serial exception catching and automatic reconnection.

2. MCU Firmware (V3F Core):
   - Configured TIM2 CH1/CH2/CH3 for 50Hz PWM output with 2us pulse width resolution for Pitch, Yaw, and Roll servos.
   - Refactored UART routing (uart_router.c) using WCH standard peripheral library: USART3 (115200bps) for face coordinates, USART4 (9600bps) for gestures, USART2 (9600bps) for voice command reception, USART1 (19200bps) for voice broadcast.
   - Addressed official board hardware NC resistor open circuit by software remapping PD5/PD6 to USART2.
   - Implemented a robust "overwrite single-buffer" RX interrupt mechanism on USART3 to prevent ORE (overrun) errors.
   - Developed a "three-segment speed approximation" algorithm (servo.c) to replace traditional PID, achieving smooth logarithmic convergence without overshoot.
   - Implemented a Roll axis dynamic polynomial differential compensation algorithm based on Pitch/Yaw offsets to keep the horizon stable.
   - Built a 7-line diagnostic telemetry dashboard on ST7735 TFT display (SPI2) for real-time monitoring of servo angles, gesture codes, and serial packets.

3. MCU Firmware (V5F Core):
   - Adapted servo.c/servo.h and main.c for V5F peripheral coordination.

4. Repository configuration:
   - Configured .gitignore to exclude build artifacts and IDE folders (.idea, .mrs, word/).

+++ V5F/App/servo.h
@@ -13,8 +13,8 @@
 /* 舵机PWM初始化 (TIM2 CH1/CH2/CH3, 50Hz) */
 void Servo_PWM_Init(void);
 
-/* 设置舵机角度 (0-180°) */
-void Servo_Set_Angle(Servo_Channel ch, uint8_t angle);
+/* 设置舵机角度 (0-180.0°) */
+void Servo_Set_Angle(Servo_Channel ch, float angle);
 
 /* 平滑移动舵机 */
 void smooth_move_servos(void);
@@ -23,7 +23,7 @@ void smooth_move_servos(void);
 void V5F_ProcessSharedData(void);
 
 /* 全局变量 */
-extern int current_angle[SERVO_COUNT];
-extern int target_angle[SERVO_COUNT];
+extern float current_angle[SERVO_COUNT];
+extern float target_angle[SERVO_COUNT];
 
 #endif /* __SERVO_H */
